changed duration of FF from int to int65; should give us enough head room

+++ cpp_dll_visual_studio/FF_UWP_WIN32_dll/FF_UWP_WIN32_dll/FF_WIN32_dll.cpp
@@ -56,7 +56,7 @@ bool initRacingWheel()
 
 }
 
-int initForceFeedback(int samplingTime) {
+int initForceFeedback(int64 samplingTime) {
     TimeSpan time;
     time.Duration = samplingTime; // samplingtime is in 100 ns Ticks --> Ts*1e7 Ticks
